+++ examples/simulation-study/src/tree-figure.R
@@ -1,0 +1,9 @@
+library(ape)
+library(ggplot2)
+library(ggtree)
+
+reconstructed_tree <- read.newick("demo-newick-string-recontree.txt")
+
+g <- ggtree(reconstructed_tree)
+
+ggsave("out/reconstructed-tree.pdf", g)